requirements.txt
Click==7.0
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
SQLAlchemy==1.3.22
Werkzeug==0.16.0
Flask-Uploads==0.2.1
pandas==1.1.4
numpy==1.19.4
openpyxl==3.0.5
csvtemplate==0.1.0
simple_salesforce.login==0.1.0
phonenumbers==8.12.24
simple-salesforce==1.11.0
datetime==4.3
os==0.1.0
werkzeug.utils==0.16.0
encodings==0.1.0
salesforce_bulk==1.0.0
requests==2.25.1